nimble/controller: Initiate data length update if remote supports it
We initiate data length update procedure as soon as we receive features
from remote and know that this is supported on connection.

This fixes issue with some crippled controllers which do not indicate
DLE support but will reply with LL_LENGTH_RSP but with broken payload
(e.g. the one in Sony Xperia Z5). This would result in LL timeout since
we properly ignore broken PDU.