sys/baselibc: Fix vsnprintf with size=0
In case of size=0, vsnprintf should not write anything to buffer - not
even a null-termination character (and then it's safe to pass NULL as a
buffer). This is defined in C99 standard.

Here, we allow to write up to size bytes into the buffer but then need
to null-terminate string properly even in case size-bytes were written
to buffer.