MYNEWT-813 RTT console i/p limited to 16 chrs
- Segger RTT takes input of 16 max based on BUFFER_SIZE_DOWN defined in
hw/drivers/rtt/include/rtt/SEGGER_RTT_Conf.h. This should be made a
syscfg. I was seeing issues with Thingy with jLinkExe version 6.00c, where I
could not issue sensor read command on it, based on the echo it seems
that the code just takes 16 characters of input.
Fix is to define MYNEWT_VAL(RTT_BUFFER_SIZE_DOWN) in syscfg.yml and set
it by default to a higher number if somebody would like to get it
working with jLinkExe version 6.00c.