oic: Per-resource transport layer sec requirements
Each resource can be configured to require transport layer encryption,
transport layer authentication, neither, or both.  When a peer accesses
the resource, the OIC server verifies the security properties of the
underlying connection.  If the minimum security requirements are not
met, the server responds with a 403 forbidden error.

This form of security is an alternative to the "coaps" DTLS security
that is defined in the CoAP spec.