BLE Host - Auto own-addr-type values
This commit introduces two new own-addr-type values, usable only by the
host:

    #define BLE_OWN_ADDR_AUTO_PUB                (0xf0)
    #define BLE_OWN_ADDR_AUTO_PRIV               (0xf1)

These values indicate that the host should choose an appropriate
own-addr-type value, depending on what is available.  The procedure for
choosing an address type is as follows:

    if privacy requested:
        if we have a random static address:
             --> RPA with static random ID
        else
             --> RPA with public ID
        end
    else
        if we have a random static address:
             --> random static address
        else
             --> public address
        end
    end

The motivation for this change is to allow BLE packages to be more
self-contained.  If a package doesn't care what type of address it uses,
it can just use whatever is available.