nimble/host: Refactor HCI commands sending functions
This refactors the way HCI commands are constructed and pass to
controller. Instead of embedding opcode and len in command parameters
buffer just pass it along with it. This has few adventages:
 - opcode OCF/OGF are contructed once on build time
 - buffer is always pass along with len
 - code is less confusing for static analysis tools (0 len buffer is
   always pass as NULL so no dangling pointers)
 - it is easier to follow (and debug)

Nice bonus is ROM size reduction of net_nimble_host.a (54304->53710).

This is also first step to making use of packed structures for HCI
command instead of builders functions. This should reduce code size
even further.