nimble/host: Allow to hide GATT services from clients
This patch introduces ble_gatts_svc_set_visibility() API to allow any
service to be hidden (and then restored) from clients. From GATT pov,
the service is still registered and has the same handle range assigned
and it is ATT server which hides the attributes from the client.

To achieve this, there is also new API in ATT server which allows to
remove range of attributes from the server. These attributes are not
removed entirely, but moved to separate list and can be restored later
in unchanged state.

Since it is perfectly fine to hide/restore service during an active
connection, it is up to application to inform clients about the change.