nordic: LF clock modifications
The following changes were made to the code regarding the LF
clock for nordic platforms. The initial impetus for this change
was that there was no support for using a synthesized 32kHz clock
for nrf52 platforms. LF clock management was different between
the nrf51 and nrf52 as well and this commit is intended to make
them consistent. Here is a list of the changes made:

1) Always turn on the LF clock for nordic bsps/platforms.
2) Move the LF clock source syscfg definitions from the BSP into
the mcu.
3) BSP now sets the proper LF clock based on the BSP hw (either
external, synthesized or internal RC oscillator).
4) The OS will now always use the 32768 crystal for OS time. All
code that dealt with it using a different clock source has been
removed.
5) Both the nrf51 and nrf52 now use the same RTC time resources.