net/lora: Additional modifications
A few additional changes with this commit. The first was to
disable the sx1276 spi before configuration. This should not
be needed but in case the spi was enabled prior to running
this image this make sure to disable it first. The second change
was to check for the lora timer number to be -1 (and not 0).
This insures that a timer number was set in the syscfg. The user
must choose a valid timer number. The last change was to make sure
that we check for additional enqueued transmissions after we are
finished with one.