nimble/phy: Fix possible race between wfr and EVENTS_ADDRESS
See comment in code for details.

This was not observed with old wfr code since wfr calculation for TX-RX
was less accurate and most importantly it used 16 usecs (jitter only
applicable to anchor point timing) to adjust for clock accuracy instead
of 2 usecs (as defined by Core spec) so the wfr time was always couple
of usecs after EVENTS_ADDRESS.

With new wfr code the calculations are more accurate so it's possible
both events will happen at the same time.