Move SPI and I2C config definitions from BSPs to MCUs.
Also add restrictions between I2C and SPI on nRF5x boards, as was already done for the EE-02.

This is primarily a refactoring, but it could cause some builds to fail.  In particular, any BSP not the core that defines these settings will see “setting XXX redefined” errors.  And any nRF5x app or target that had both I2C_0 and SPI_0 enabled (or _1) will run into the new restrictions.

Otherwise, all settings (interfaces enabled, pin numbers) are unchanged, with one exception:  The VBLUno boards had enabled I2C and SPI 0 and 1.  To resolve the conflict, I disabled SPI, as at least I2C_0 seems like it ought to be enabled by default for the on-board OLED.