native_sockets: Poll write no matter read state.
[Prior to this commit]

The native_sockets task periodically does two jobs:

1. Polls its sockets for incoming data.
2. Checks if any sockets have unsent data.  If so, it attempts to send
it.

There is a bug here - if job 1 determines that there is no incoming
data, job 2 is never performed.  If there is any unsent data, it won't
get sent job 1 detects some incoming data to receive.  The bug was
caused by an early `continue;` when a call to `poll()` indicated no
sockets ready to receive.

[Changes]

Now, the sockets are polled for reads and writes (previously they were
only polled for reads).  The task loop always checks for read and write
events on all sockets.