nimble/ll: Fix scheduling of extended advertising events
The advertising state machine uses flag to distinguish between two phys
when handling events sent from interrupt to ll task, but is problematic
since flag is set in an interrupt and then usually checked in ll task
so race may occur when two events for different phys are passed to the
task. This makes advertising unstable in non-ideal conditons.

To fix this we use separate events for pri and sec phys for advertising
so there will be no races when accessing the flag since it is done in
isr only (except for cases where we want to stop advertising but it
does not matter there).