net/lora: Class C unconfirmed transmission bug fixes
This commit fixes some issues with unconfirmed transmissions for
class C devices. The first issue was that unconfirmed packets
were not getting freed of the transmit queue because the mcps
confirm was not being called. We now make sure if we receive a
frame after an unconfirmed transmission that we call the mcps
confirm. The other fix was to make sure we receive on window 2
if we send an unconfirmed transmission and we timeout after
waiting to detect start of frame.