mesh: Fix and clean up iv update procedure
This patch fixes issue when receiving iv index greater than current
index + 42 in update mode. According to Specification when node is in update
state it should only accept iv index equal to the current iv index. When
node is in normal mode it should ignore index that is greater than current
index + 42.

This allows to pass MESH/NODE/IVU/BI-02-C.

Also this patch cleans up the iv update procedure, to make it easier to
read.