nimble/phy: Fix scheduling gap between coded and uncoded
TX scheduled at certain time will be off by 2 usecs depending on which
phy (coded/uncoded) it is scheduled. This is due to different T_txdelay
on coded and uncoded. It's not a big problem when scheduling e.g. conn
events, but it can affect ext advertising if primary and secondary use
different phys.

Right now, if we use coded for primary, aux pdu is calculated at either
1.77 or 1.80 msecs from 2nd ext adv pdu which is also multiply of full
ticks (57.99 or 58.98 ticks respectively) without usecs remainder. Not
a problem if secondary is also coded, but in case it is uncoded aux is
sent 2 usecs too early since T_txdelay is 2 usecs less for coded that
for uncoded. To fix this we need to take T_txdelay into account when
scheduling TX to start in the future.