Removing use of undefined TASKS_START behavior
This is a follow up commit to a the same fix in the nRF52. Tested using the bleprph app - but only on an emulator. Seems to be working the same as before. Happy to help testing further!

The problem was that the behavior of the RNG TASKS_START is defined to be write only. Reading from it is undefined and though it might have worked up until now doesn't mean it'll constantly work. I was only able to catch on to this using an emulator.