BLE Host - Delay sync callback after IRK restore
-Prior to commit-

On startup, the host executes the following steps:

1. HCI startup sequence
2. Call sync callback
3. Restore IRKs from flash

The problem is that step 3 (IRK restoration) causes a GAP preemption if
there are any persisted IRKs.  If the application initiates a GAP
procedure during the sync callback, it will be immediately preempted.
This happens in the bleprph app, for example.

-After commit-

Don't call the sync callback until after IRK restoration is complete.