BLE HCI transport - emspi
This is an HCI transport that works over SPI.  It uses EM's SPI
protocol, as described in the EM9304 data sheet:
http://www.emmicroelectronic.com/sites/default/files/public/products/datasheets/9304-ds.pdf

Currently only the host side is supported.

In addition, this package only builds when the MCU is an APOLLO 2.  This
is due to a dependency on the ambiq SDK.