apollo2 - Ensure OS tick timer is running.
The code assumes that when it does not reschedule the OS tick timer
(STIMER), it will expire on the next OS tick.  This assumption failed on
the first tick after startup, as the ISR hadn't run yet.  Consequently,
the idle task would never terminate if a scheduling event occurred on
the first tick.

The fix is to set the STIMER to expire on the next OS tick at init time.