sys/console: Fix console freeze
console_queue_char() may be called from interrupt context
and waiting for uart to drain the buffer may cause the console to
freeze. In these situations we should drop the character if
the buffer is full.

This is an easy solution to this problem, but we should redesign
console to not do this kind of processing in interrupt context.