mempool - Extended mempools
This commit adds a second type of mempool: extended.  An extended
mempool supports basic mempool functionality, with the following
additional feature:

    * put callback

For mempools configured with this callback, `os_memblock_put()` executes
the callback instead of freeing the specified block.  It is then the
callback's responsibility to free the block to its pool via the new
function `os_memblock_put_no_ext()`.