kernel/os_mempool: Fix for incorrect check on os_mempool_init
This is regression from abd8c4598b as it was changing API incorretly.
Documentation says that 'block_size' parameter in os_mempool_init() is
"Minimum size of an individual element in pool"

It does not required from the user to take into account of 'block_size'
an internal overhead which is sizeof(struct os_memblock).

The only requirement on user is to make sure that provided 'membuf' has
enough bytes which can be calculated using OS_MEMPOOL_BYTES() macro.

Without this patch we have assert when creating mempool with block_size=4