apps/btshell: Allow to append some extra data to adv/scanrsp data
Optional 'extra_data_len' parameter for both advertise-set-adv-data and
advertise-set-scan-rsp will append specified number of bytes to adv or
scanrsp data respectively.

Appended data are not valid advertising data. Instead, there is counter
encoded as BCD on every 2 bytes, so we have sth like this:
00:00:00:02:00:04:00:06:00:08:00:10:00:12:00:14:...:16:48:16:50
This allows to easily check whether data are continuous and properly
fragmented in the air which is especially useful for testing extended
advertising.