nimble/ll: Fix generating RPA when advertising
When starting SM, we always overwrite AdvA and InitA to "identity"
values and then we should overwrite them again with RPAs, if necessary.
However, since we call function which rotates RPA for advertising SM it
will not update addresses if called before rotate timeout fired.

This means that 1st time we start SM addresses are set properly, but
then we may use identity addresses instead of RPAs if SM is restarted.

This patch fixes this by ensuring proper RPAs are always configured in
SM.