nimble/ll: Add note about InitA when initiating connection with privacy
Even with privacy enabled we have to use our identity address as InitA
when peer is not on our resolving list. Since this is a bit counter-
intuitive, let's add a note with spec reference so we do not need to
waste time in future figuring out why it is implemented like this.