apps/ble: Fix GATT discovery in some apps
This fixes an issue with GATT discovery in few sample apps which
results in an infinite loop with characteristics discovery if peer's
database has service with no characteristics due to invalid check for
empty service.

The same fix is applied to bletiny and btshell which do not have this
issue anymore, but the original fix just hidden the issue instead of
fixing the root cause.