hw/drivers/sx1272: Make driver shield compatible and fix driver issues.
The main changes in this commit were to make all GPIO that interface
the chip to the MCU MYNEWT_VALs so they can be configured either
in a target or bsp. This makes creating a bsp + shield combo
simple by defining the interface GPIO in a target.

Some additional fixes/changes are in this commit such as:
1) Using the proper reset sequence.
2) Fixing a bug with the spi/read write command.
3) Adding a MYNEWT_VAL to select a separate spi routine to be
used to read/write buffers to the chip. The blocking,
byte-by-byte hal spi api is not working on the 52840
when combined with the sx1272 for some reason so this is a
temporary work-around.