Add option do configure spi master/slave number
This adds new syscfg options to pass the number of the slave/master
peripheral to use in the tests. This customization is required when
testing on boards that have SPI0 pins already used for other purposes or
not easily reachable in the conn headers.

This also allows for using a single board doing spi testing between two
different spi peripherals.