net/lora: Update sx1272 to latest semtech code.
Decided to bring this file up to date with latest code in semtech
repo. Changes were all coding style except for the setting of
FIFO address pointer after a received frame. While this change
may not be needed given how we access the FIFO it is more correct
to have it than not.