boot_serial - Allow UART string for serial mode.
If enabled, the serial boot loader listens on the UART for a string
for a configurable amount of time.  If the string is detected, the boot
loader enters serial mode.  If the string is not detected when the
timeout expires, the boot loader proceeds in normal (non-serial) mode.

This feature is intended for setups that lack a suitable pin to use for
enabling serial mode (either the pins are inaccessible or inconvenient
to use).

Enabling this feature added 248 bytes to the boot loader binary when
built on the nRF52dk with the default string setting ("nmgr").