sys/log: Add wrapper for logging to slot1
This is thin wrapper for logging to FCB located in slot 1. The
difference between regular FCB handler is that it is possible for other
code (e.g. imgmgr) to lock slot 1 before loading new image there and
log handler will reject adding new log entries so image data is not
corrupted.