hw/drivers/lora: Add MYNEWT_VALs to select PA_BOOST or RFO
The sx1272 and 1276 have different possible RF output paths. The
code was assuming that PABOOST was always used for the sx1272
and that PABOOST was used for the HF band on the sx1276 always.

The new code allows the bsp/app/target/etc to specify the RF
output path for transmit. The added configuration variables are:
SX1272_USE_PA_BOOST and SX1276_HF_USE_PA_BOOST and
SX1276_LF_USE_PA_BOOST. The default is to use the normal RF
output path in all cases.