sys/log: Add log entry type
This patch extends log header with 'type' field where type of data
stored in entry can be put.

Until now it was assumed that log entry is a short string, but since
it can be as long as 64k it makes sense to allow to put some binary
data there (like CBOR). This allows to handle data from log entry in
a better way, e.g. console handler can pretty-print log entry.

New functionality is only available if LOG_VERSION is set to 3. To
maintain backwards compatibility, default log version is set to 2
and only string type is allowed.