apps/blemesh_light: Add demo app that uses Light Lightness model
This app serves as an example of using Generic models to
implement a Lighting app. You can control it from Silicon Labs'
Android Mesh app. Configure it the device as Level model and use slider
to control the lightness. You can use LEDs on board controlled with PWMs
or use NEOPIXEL LED matrix.

This app uses a special API that was exposed for demo purposes.