kernel/os - tu_restart on task terminate.
Normally, a task handler should never return.  The situation is a bit
different for unit tests that run in the OS.  Typically, such tests are
short-lived and consist of only two tasks: 1) the default task, and 2)
the "test task", which runs the actual tests via a sequence of function
calls and assertions.  For these tests, it is convenient to signify test
completion by allowing the test task to simply return, rather than
explicitly completing the test via a call to `tu_restart()`.

This commit changes the scheduler code in `SELFTEST` mode such that
`tu_restart()` is automatically called when a task handler returns.