testutil - TEST_CASE_TASK() for in-OS tests
Most tests don't utilize the OS scheduler; they simply run in main(),
outside the context of a task.  However, sometimes the OS is required to
fully test a package, e.g., to verify timeouts and other timed events.

This commit simplifies the implementation of test cases that require the
OS.  This is done by the addition of a macro: `TEST_CASE_TASK()`.  This
macro is identical in usage to `TEST_CASE()`, except the test case it
defines performs some additional preliminary work:

1. Calls `sysinit()`
2. Creates the default task.
3. Creates the "test task" (the task where the test itself runs).
4. Starts the OS.

The `TEST_CASE_TASK()` function body is what actually runs in the test
task.

This eliminates the need to define two tasks, two stacks, and two
handlers for each test case.