sys/log: Make sure to output at least single log entry via nmgr
There's some magic limit of 400 bytes when writing response to nmgr.
While it works with string entries, which "by design" are limited to
128 bytes so we will always fit at least single log entry, it may fail
with CBOR or binary log entries which can easily exceed this value. As
a result it may be impossible to include at least one entry in response
so whole log becomes virtually unreadable.

This patch makes sure that at least single entry is always returned in
response so even with long log entries we can still read log.