kernel/os: Add helpers for converting ms<->ticks
These helpers do not check for overflow so should be used with care,
but in general they allow for more compact code.

Note that would be nice to have them as static inline but since we need
OS_TICKS_PER_SEC in header file, there are some weird dependencies in
os/os_*.h files which needs to be resolved first so code can build
properly.