sysview: Refactor os_trace APIs
This makes several improvements in os_trace APIs described below.
Note that these APIs are used to trace Mynewt's internals so renaming
them should not be a problem for external users. The only potential
backwards incompatibility is renaming of ISR enter/exit calls but this
can only occur in external BSPs and drivers (assuming anyone actually
uses this in own BSPs or drivers) and is just a simple rename.

Changes made:

* Make all APIs "static inline" to avoid unnecessary calls

* Improve API names consistency:
  - os_trace_api -> anything related do API calls
  - os_trace_isr -> anything related to ISR calls
  - os_trace_task -> anything related to task activity

* Use "struct os_task" pointer as task id instead of an actual task_id
  assigned by Mynewt. This is to be able to easily obtain task struct
  from SystemView since tasks are shown by name there anyway

* Remove unused (and unlikely to be used) APIs