apps/btshell: Add cmd_set_scan_opts command
This patch extends btshell with cmd_set_scan_opts command, that
limit the data printed out to the serial console during
scanning. It has been noticed that the large amount of data
sent on UART might cause the LE Extended Adverising packets
to be truncated.

The command params are:
* ignore_legacy - to not print legacy advertisements
* decode_limit - to limit the data to be printed out to serial
console.