sysview: Add option to use hal_timer for timestamping
Without built-in cycle counter we need to use different source of
high-resolution timestamp to have accurate trace and we can use one of
hardware timers for this purpose (via hal_timer).

This patch adds an option to configure one of hardware timers as a
source of timestamp of selected frequency (default: 1 MHz) and thus
enables accurate timestamps on e.g. Cortex-M0.

Note that if possible, built-in cycle counter is still the best option
to use here.