util/mem: mem_pullup_obj() utility function.
Add the `mem_pullup_obj()` utility function.  This function applies a
pullup operation to the supplied mbuf and returns a pointer to the start
of the mbuf data.  This is simply a convenience function which allows
the user to access the mbuf data without a cast.  On failure, the
provided mbuf is freed.