lis2dw12: Fix self testing function error
Per datasheet configuration that was supposed to switch on
positive self test mode, did it, but also turned off register
auto increment functionality.
This effectively disabled possibility to read
normal values from accelerometer since those reading are
done in blocks and not single registers.
After the self test, original configuration was meant to be restored,
alas it used multi register write that was just turned off,
so the configuration was not restored.

Now IF_ADD_INC is set in per-datasheet configuration.