boot/split: Ensure conf loaded before status read
This removes the need for a manual call to `conf_load()` in `main()`,
prior to the call to `split_app_go()`

`split_app_go()` is the function which decides whether to boot into the
second app or to proceed with the current one).  This function requires
that the configuration has been loaded, as the action it takes depends
on the `split/status` config setting.