sys/log/full: Add "read_body" functions to API
This fixes #1173.

This commit adds three functions to the log API:
* `log_read_hdr`
* `log_read_body`
* `log_read_mbuf_body`

`log_read_hdr` just reads the specified entry header from the log.

The to "read body" functions are very similar to the existing "read"
function.  The only difference is that the specified offset starts from
the beginning of the message body, not the from the beginning of the
entry.  That is, the user doesn't need to add `LOG_ENTRY_HDR_SIZE` to
the offset that gets passed in.