rtt: Move RTT control block to separate section
This patch puts RTT control block to separate ".rtt" section. This
allows to put it at stable location in linker script so it does not
move after rebuilding firmware so there is usually no need to restart
RTT clients after loading new firmware.

However, since this section is not automatically zeroed (because this
would require changing every single BSP) it needs to be manually zeroed
as early as possible. This is done in os_init() since package init may
be called to late - in case SystemView is enabled it is possible for
RTT code to be called before we reach main().