net/lora: Fix lora driver spi access
The code did not display receive interrupts from occurring when
an attempt to transmit was made. This could cause a race condition
where the code is loading the semtech radio FIFO for transmit and
then gets a receive packet interrupt which interrupts the current
spi access and could cause a spi driver lockup (which results in
a watchdog).

These changes should address that issue by disabling receive
interrupts and clearing any receive events prior to transmitting.