rtt: Update to rev 11303
This updates RTT code to rev 11303.

There are also changes in the way this is integrated to make updates
easier in future:
- SEGGER_RTT_Conf_Mynewt.h is removed (no need for it)
- SEGGER_RTT_Conf.h is replaced with custom file which has setup that
  applies to Mynewt only (other settings are initialized to default
  values in RTT code anyway)
- package init code is in separate file

This means in future we can overwrite everything in .c and .h files to
upgrade (except for configuration, which we just keep).

Also, with proper configuration file locking functions work properly
since previously they were overwritten by built-in functions from
default config file and were not working properly in Mynewt.