mgmt/imgr: Add support to resume image upload
1st chunk of data can include "datahash" field which is simply SHA256
computer over all data to be uploaded. This hash is stored in upload
state so in case transfer was interrupted we can detect this on next
attempt and simply resume transfer at previous location by replying
with latest offset from upload state.

"datahash" field is not mandatory so if it is missing we'll just start
new upload as usual.