kernel/os: Preempt when high priority task created
Before this change, a newly-created high-priority task would not
immediately preempt the current one.  Preemption would not occur until
the next call to `os_sched()` (e.g., via a mutex unlock).

This commit causes `os_sched()` to be called whenever a new task is
created (so long as the OS has been started).