mgmt/imgmgr: Clean up upload request handler
This commit partitions the image upload request handler into two parts:
    1. read-only "inspection" phase
    2. "action" phase

The inspection phase verifies the values in the request and determines
which actions to take.

The action phase performs the following actions, as needed:
    1. Set global variables.
    2. Lock image slot 1.
    3. Erase image slot 1.
    4. Write image data to slot 1.

Rationale for this commit: I needed to add a custom "hook" that allows
the application to reject an upload request.  I wanted the callback to
get called only after the request handler determined that it would
accept the request, but before any state changed as a result of
receiving the request.  I wasn't able to meet these requirements with
the old structure of the request handler, hence this commit.