mgmt/imgmgr: Allow app to reject image upload reqs
Allow a callback to configured in the imgmgr module.  If one is
configured, the callback is called immediately before the request is
processed.  The callback has the following properties:

The callback's return code determines whether the upload request is
accepted or rejected.  If the callback returns 0, processing of the
upload request proceeds.  If the callback returns nonzero, the request
is rejected with a response containing an `rc` value equal to the return
code.

@param offset                The offset specified by the incoming
                                request.
@param size                  The total size of the image being uploaded.
@param arg                   Optional argument specified when the
                                callback was configured.

@return                      0 if the upload request should be accepted;
                             nonzero to reject the request with the
                                specified status.

typedef int imgr_upload_fn(uint32_t offset, uint32_t size, void *arg);