battery/adc: Add activation pin support
Battery voltage reading often uses resistor divider that is
actiaved with GPIO pin for energy saving (divider is not powered
when not needed).
This change allows to specify GPIO pin that must be activated
before ADC can be used.
GPIO level also can be specified allowing divider activation with 0 and 1.