Add support for the Bosch BME680 Environmental Sensor
This adds support for the BME680 sensor to the mynewt sensor framework.
The driver supports both I2C and SPI modes (NOTE: SPI not tested at time
of commit). The syscfg file can be used to configure whether or not
floating point compensation is used or not, and whether the output values
use the standard mynewt sensor value types (found in temperature.h,
pressure.h and humidity.h) or the Bosch definitions (using struct bme680_field_data
in bme680_defs.h). The gas resistance output is either a float or a uin32_t,
depending on whether floating point compensation is turned on or not.

The sensor can be used standalone with this driver, but is ultimately
designed to be used alongside the Bosch BSEC sensor fusion library for
best performance (https://www.bosch-sensortec.com/bst/products/all_products/bsec).
A mynewt module adding support for this library will be added at a later date.

The sensor_creator code has been modified to include an off-board instance of the
sensor using I2C0. This can be enabled usign the BME680_OFB syscfg value.